pyproject.toml
[project]
name = "trivia-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asgiref==3.8.1",
    "certifi==2025.4.26",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "cryptography==45.0.3",
    "defusedxml==0.7.1",
    "django==5.2.1",
    "django-mysql==4.17.0",
    "django-redis==5.4.0",
    "django-rest-framework==0.1.0",
    "djangorestframework==3.16.0",
    "djangorestframework-simplejwt==5.5.0",
    "dotenv==0.9.9",
    "h11==0.16.0",
    "idna==3.10",
    "mysqlclient==2.2.7",
    "oauthlib==3.2.2",
    "pillow>=11.2.1",
    "pycparser==2.22",
    "pyjwt==2.9.0",
    "pymysql==1.1.1",
    "python-dotenv==1.1.0",
    "python3-openid==3.2.0",
    "redis==6.1.0",
    "requests==2.32.3",
    "requests-oauthlib==2.0.0",
    "social-auth-app-django==5.4.3",
    "social-auth-core==4.6.1",
    "sqlparse==0.5.3",
    "urllib3==2.4.0",
    "uvicorn==0.34.2",
]
